Refactor codebase for improved clarity
