+++ db/db.go
@@ -10,41 +10,32 @@ import (
 	_ "github.com/lib/pq"
 )
 
-var DB *sql.DB
-
-func InitDB() (*sql.DB, error) {
+func Connect() (*sql.DB, error) {
 	err := godotenv.Load(".env")
 	if err != nil {
 		log.Fatalf("Error loading .env file")
 	}
 
-	connStr := fmt.Sprintf("user=%s dbname=%s password=%s sslmode=disable", GetEnvVariable("DATABASE_USER"), GetEnvVariable("DATABASE_NAME"), GetEnvVariable("DATABASE_PASSWORD"))
+	connStr := fmt.Sprintf("user=%s dbname=%s password=%s sslmode=disable", getEnvVariable("DATABASE_USER"), getEnvVariable("DATABASE_NAME"), getEnvVariable("DATABASE_PASSWORD"))
 
-	DB, err = sql.Open("postgres", connStr)
+	db, err := sql.Open("postgres", connStr)
 	if err != nil {
 		return nil, fmt.Errorf("error opening database connection: %v", err)
 	}
 
-	err = DB.Ping()
+	err = db.Ping()
 	if err != nil {
 		return nil, fmt.Errorf("error pinging database: %v", err)
 	}
 
-	return DB, nil
+	return db, nil
 }
 
-func GetEnvVariable(key string) string {
+func getEnvVariable(key string) string {
 	envVariable, ok := os.LookupEnv(key)
 	if !ok {
 		log.Fatalf("Could not establish database connection: no %s env var", key)
 	}
 
 	return envVariable
-}
-
-func CloseDB() {
-	if DB != nil {
-		DB.Close()
-		log.Println("Database connection closed")
-	}
 }
